+++ pypepper/network/http/__init__.py
@@ -1,0 +1,7 @@
+"""FastAPI HTTP server helpers."""
+
+from . import server
+
+__all__ = [
+    "server",
+]